backend/requirements/dev.txt
-r base.txt  # includes travis.txt requirements file

colorama==0.3.9
django-coverage==1.2.4
django-debug-toolbar==1.8
django-extensions==1.9.1
django-cors-headers==2.1.0
factory-boy==2.9.2
ipdb==0.10.3
pdbpp==0.9.1

# tests
astroid==1.5.3
codecov==2.0.9
coverage==4.4.1
django-nose==1.4.5
pycodestyle==2.3.1
pylint==1.7.2
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
